begin split of dataloader
